+++ src/main/java/com/jig/blog/repository/BoardRepository.java
@@ -17,7 +17,7 @@ import java.util.Optional;
  */
 
 // @Repository // JpaRepository를 상속하면 bean 생성을 해주는 어노테이션이 없어도 자동으로 bean 등록이 된다.
-public interface BoardRepository extends JpaRepository<Board, Integer> {
+public interface BoardRepository extends JpaRepository<Board, Long> {
 
     // #1. N+1 문제 해결을 위해 EntityGraph 또는 fetch join 사용
 /*
@@ -33,7 +33,7 @@ public interface BoardRepository extends JpaRepository<Board, Integer> {
     Page<Board> findAllWithUserBy(Pageable pageable);
 
     @Query("select s from Board s join fetch s.user where s.id = :id")
-    Optional<Board> findWithUserById(int id);
+    Optional<Board> findWithUserById(Long id);
 
 //    // fetch join을 사용하여 board와 reply를 함께 조회하는 메서드
 //    @Query("SELECT b FROM Board b JOIN FETCH b.replies WHERE b.id = :boardId")
@@ -42,5 +42,5 @@ public interface BoardRepository extends JpaRepository<Board, Integer> {
     // todo: @Modifying 관련 추가 공부 필요할 듯
     @Modifying // 데이터를 벌크로 수정하는 임의 작성 jpql에 붙여주는 어노테이션이고 영속성 관리에 대한 옵션 설정이 필요 할 수 있음
     @Query("delete from Board b where b.user.id = :userId")
-    int deleteAllByUserId(@Param("userId") int userId);
+    int deleteAllByUserId(@Param("userId") Long userId);
 }
